Get stream depletion and changes in water level from pumping

This function estimates stream depletion fraction (using `get_stream_depletion_fraction`) and changes in water level at an observation well (`get_aquifer_drawdown_ratio`) due to abstraction from a pumping well at time `t` after pumping initiates. Like `get_stream_depletion_fraction`, and unlike `get_aquifer_drawdown_ratio`, The function accounts for a the effect of the stream as a constant head boundary. See Glover (1954).

get_depletion_from_pumping(
  df,
  x1 = NULL,
  x2 = NULL,
  y = NULL,
  K = NULL,
  D = NULL,
  V = NULL,
  t = NULL,
  well_diam = NULL
)

Arguments

df

data.frame with columns specifying all parameters

x1

Distance between well and river

x2

Distance of observation well to stream

y

Distance between pumping and observation well (parallel to stream)

K

Saturated hydraulic conductivity

D

Depth of aquifer

V

Drainable porosity of aquifer

t

Time from pumping onset at which to calculate stream depletion fraction

well_diam

Diameter of the well, inside which drawdown does not increase. Defaults to 0.

Value

A `data.frame` with two columns: `stream_depletion_fraction` and `aquifer_drawdown_fraction`. To calculate stream depletion and changes in water level, multiply these values by the pumping rate.

Details

Get stream depletion and changes in water level from pumping

Examples

library(units)
x1 <- set_units(c(1, 5, 10) * 1e3, "ft")
x2 <- set_units(1e3, "ft")
y <- set_units(1e3, "ft")
D <- set_units(100, "ft")
K <- set_units(0.001, "ft/sec")
t <- set_units(5, "year")
V <- 0.2 # unitless
depletion_from_pumping <- get_depletion_from_pumping(x1 = x1,
                                                     x2 = x2,
                                                     y = y,
                                                     K = K,
                                                     D = D,
                                                     V = V,
                                                     t = t)
depletion_from_pumping
#>   stream_depletion_fraction aquifer_drawdown_ratio
#> 1                 0.9365474    -1.2707109 [s/ft^2]
#> 2                 0.6905933    -0.5705381 [s/ft^2]
#> 3                 0.4259739    -0.2299550 [s/ft^2]

# Specifying parameters as named data.frame columns
library(tibble) # simplifies specifying data.frames with units objects
df <- tibble(x1 = x1, x2 = x2, y = y, K = K, D = D, V = V, t = t)
depletion_from_pumping <- get_depletion_from_pumping(df)
depletion_from_pumping
#>   stream_depletion_fraction aquifer_drawdown_ratio
#> 1                 0.9365474    -1.2707109 [s/ft^2]
#> 2                 0.6905933    -0.5705381 [s/ft^2]
#> 3                 0.4259739    -0.2299550 [s/ft^2]